move packet-building functions to dlm-util
