+++ Core/Inc/dlm-transmit_ram_data.h
@@ -14,7 +14,5 @@
 
 void transmit_ram_data_init(DATA_INFO_NODE* ram_ptr);
 void transmit_data(UART_HandleTypeDef* uart);
-U8 packetize_node(DATA_INFO_NODE* node, U8 packet[]);
-U8 append_byte(U8 packet[], U8 packetLength, U8 byte);
 
 #endif /* INC_DLM_TRANSMIT_RAM_DATA_H_ */
