+++ Core/Src/dlm-move_ram_data_to_storage.c
@@ -16,6 +16,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include "dlm-util.h"
 
 
 // This is the same head for the RAM storage linked list in manage_data_aquisition
@@ -214,52 +215,6 @@ void build_data_string(U8 data_str[], DATA_INFO_NODE* data_node)
     {
         data_str[c + STORAGE_PARAM_SIZE + TIMESTAMP_SIZE] = (U8)(data_union.u64 >> (((DATA_SIZE - 1) - c) * BITS_IN_BYTE));
     }
-}
-
-
-// convert_data_to_dpf
-//  Function to take in a data node, get the data stored in it, and return
-//  the double precision float representation of that value to be stored on
-//  the external SD card
-double convert_data_to_dpf(DATA_INFO_NODE* data_node)
-{
-    // switch to get the data out of the data_node
-    switch (parameter_data_types[data_node->param])
-	{
-	case UNSIGNED8:
-		return (double)(((U8_DATA_NODE*)data_node)->data);
-
-	case UNSIGNED16:
-		return (double)(((U16_DATA_NODE*)data_node)->data);
-
-	case UNSIGNED32:
-		return (double)(((U32_DATA_NODE*)data_node)->data);
-
-	case UNSIGNED64:
-		return (double)(((U64_DATA_NODE*)data_node)->data);
-
-	case SIGNED8:
-		return (double)(((S8_DATA_NODE*)data_node)->data);
-
-	case SIGNED16:
-		return (double)(((S16_DATA_NODE*)data_node)->data);
-
-	case SIGNED32:
-		return (double)(((S32_DATA_NODE*)data_node)->data);
-
-	case SIGNED64:
-		return (double)(((S64_DATA_NODE*)data_node)->data);
-
-	case FLOATING:
-		return (double)(((FLOAT_DATA_NODE*)data_node)->data);
-
-	default:
-        // Something went wrong, just write 0 to data
-		return 0;
-	}
-
-    // this coude should not be reached, this is to make the compiler happy
-    return 0;
 }
