+++ Core/Src/dlm-transmit_ram_data.c
@@ -50,56 +50,3 @@ void transmit_data(UART_HandleTypeDef* uart) {
 	// remember the first node in this transmission
 	prevTransmission = transmissionHead;
 }
-
-// takes a data node, breaks it into bytes, and writes it to a packet (byte array)
-// returns the length of the packet
-U8 packetize_node(DATA_INFO_NODE* node, U8 packet[]) {
-    U8 i;
-    U8* bytes;
-    U8 packetLength = 0;
-
-    packet[packetLength] = (U8) 0x7e;
-    packetLength++;
-
-    for (i = sizeof(node->data_time); i > 0; i--) {
-        bytes = (U8*) &(node->data_time);
-        packetLength = append_byte(packet, packetLength, bytes[i - 1]);
-    }
-
-    for (i = sizeof(node->param); i > 0; i--) {
-        bytes = (U8*) &(node->param);
-        packetLength = append_byte(packet, packetLength, bytes[i - 1]);
-    }
-
-    // write the double of the data to the last 8 bytes
-    DPF_CONVERTER data_union;
-	data_union.d = convert_data_to_dpf(node);
-    for (i = sizeof(data_union.u64); i > 0; i--) {
-        bytes = (U8*) &(data_union.u64);
-        packetLength = append_byte(packet, packetLength, bytes[i - 1]);
-    }
-
-    return packetLength;
-}
-
-// escapes a byte if necessary, then appends it to the packet
-// takes the current length of the packet and the byte to append
-// returns the new packet length
-U8 append_byte(U8 packet[], U8 packetLength, U8 byte) {
-    U8 bytesFilled = 0;
-
-    // check for a control byte
-    if (byte == 0x7e || byte == 0x7d) {
-        // append an escape byte
-        packet[packetLength + bytesFilled] = 0x7d;
-        bytesFilled++;
-        // append escaped byte
-        packet[packetLength + bytesFilled] = byte ^ 0x20;
-        bytesFilled++;
-    } else {
-        packet[packetLength + bytesFilled] = byte;
-        bytesFilled++;
-    }
-
-    return packetLength + bytesFilled;
-}
